requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
ultralytics==8.3.0
opencv-python==4.9.0.80
numpy==1.26.3
pillow==10.2.0
torch==2.2.0
torchvision==0.17.0